requirements.txt
google-cloud-pubsub>=0.35.4
